+++ src/Crossbar/CrossbarCompiler.py
@@ -1,0 +1,13 @@
+class CrossbarCompiler(object):
+
+    """
+    Initialize compiler
+    """
+    def __init__(self):
+        return
+
+    """
+    Check conflicts
+    """
+    def check_conflicts(self):
+        return