+++ src/Crossbar/operation/Operation.py
@@ -1,0 +1,4 @@
+class Operation(object):
+    def __init__(self):
+        # Do somthing
+        return