+++ src/Crossbar/operation/Shuttling.py
@@ -1,0 +1,6 @@
+from Operation import Operation
+
+class Shuttling(Operation):
+    def __init__(self):
+        # Do somthing
+        return